requirements.txt
ffmpeg-python
whisper
argparse